gofault/ioc/container_test.go
package ioc

import (
	"testing"
)

type dummyService struct {
	msg string
}

func newDummyService() *dummyService {
	return &dummyService{msg: "hello"}
}

func (s *dummyService) SomeMethod() string {
	return s.msg
}

func TestContainer_RegisterAndResolve(t *testing.T) {
	c := New()
	err := c.Register(newDummyService)
	if err != nil {
		t.Fatalf("Register failed: %v", err)
	}
	svc, err := c.Resolve(&dummyService{})
	if err != nil {
		t.Fatalf("Resolve failed: %v", err)
	}
	if svc.(*dummyService).msg != "hello" {
		t.Fatalf("expected 'hello', got %s", svc.(*dummyService).msg)
	}
}

func TestContainer_SingletonScope(t *testing.T) {
	c := New()
	c.Register(newDummyService)
	a, _ := c.Resolve(&dummyService{})
	b, _ := c.Resolve(&dummyService{})
	if a != b {
		t.Fatal("singleton instances must be identical")
	}
}

func TestContainer_TransientScope(t *testing.T) {
	c := New()
	c.RegisterTransient(newDummyService)
	a, _ := c.Resolve(&dummyService{})
	b, _ := c.Resolve(&dummyService{})
	if a == b {
		t.Fatal("transient instances must be different")
	}
}

func TestContainer_ResolveUnknown(t *testing.T) {
	c := New()
	_, err := c.Resolve(&dummyService{})
	if err == nil {
		t.Fatal("expected error for unregistered type")
	}
}
